[Chore] Build Setting
