+++ Assets/AddressableAssetsData/AssetGroups/KMS.asset
@@ -80,11 +80,6 @@ MonoBehaviour:
     m_ReadOnly: 0
     m_SerializedLabels: []
     FlaggedDuringContentUpdateRestriction: 0
-  - m_GUID: d7d7599cd6f797143bf5975461644a5b
-    m_Address: M_12_Recall
-    m_ReadOnly: 0
-    m_SerializedLabels: []
-    FlaggedDuringContentUpdateRestriction: 0
   m_ReadOnly: 0
   m_Settings: {fileID: 11400000, guid: e0428a37d5604e1449d03ab9994676e9, type: 2}
   m_SchemaSet:
