--- Assets/PersonalWorkplace/LGH/Scripts/PlayerModelGenerator.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: d25bac16da093fb42a2a177c9fa64743
