+++ Assets/Plugins/Android/FirebaseApp.androidlib/res/values/google-services.xml
@@ -1,9 +1,10 @@
 <?xml version='1.0' encoding='utf-8'?>
-<resources xmlns:tools="http://schemas.android.com/tools" tools:keep="@string/gcm_defaultSenderId,@string/google_storage_bucket,@string/project_id,@string/google_api_key,@string/google_crash_reporting_api_key,@string/google_app_id">
-  <string name="gcm_defaultSenderId" translatable="false">41831429392</string>
-  <string name="google_storage_bucket" translatable="false">cheonmyungmuhwa-c8f22.firebasestorage.app</string>
-  <string name="project_id" translatable="false">cheonmyungmuhwa-c8f22</string>
-  <string name="google_api_key" translatable="false">AIzaSyCOrJ6V2XKap19JBygugCglXXf4VMQofnI</string>
-  <string name="google_crash_reporting_api_key" translatable="false">AIzaSyCOrJ6V2XKap19JBygugCglXXf4VMQofnI</string>
-  <string name="google_app_id" translatable="false">1:41831429392:android:20f1ecfd9cb251d0a2a195</string>
+<resources xmlns:tools="http://schemas.android.com/tools" tools:keep="@string/firebase_database_url,@string/gcm_defaultSenderId,@string/google_storage_bucket,@string/project_id,@string/google_api_key,@string/google_crash_reporting_api_key,@string/google_app_id">
+  <string name="firebase_database_url" translatable="false">https://cheonmyungmuhwa-d3fc4-default-rtdb.asia-southeast1.firebasedatabase.app</string>
+  <string name="gcm_defaultSenderId" translatable="false">1046450483275</string>
+  <string name="google_storage_bucket" translatable="false">cheonmyungmuhwa-d3fc4.firebasestorage.app</string>
+  <string name="project_id" translatable="false">cheonmyungmuhwa-d3fc4</string>
+  <string name="google_api_key" translatable="false">AIzaSyDfTRY8PgLXh5ORdV1h7f6EOCiPcSrM-Pg</string>
+  <string name="google_crash_reporting_api_key" translatable="false">AIzaSyDfTRY8PgLXh5ORdV1h7f6EOCiPcSrM-Pg</string>
+  <string name="google_app_id" translatable="false">1:1046450483275:android:ff94f129e1091537725dce</string>
 </resources>
