+++ Assets/Plugins/Android/GooglePlayGamesManifest 1.androidlib.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 4b4e9f011af1e4e43b28e6795f1e51b2
