+++ Assets/Plugins/Android/GooglePlayGamesManifest 1.androidlib/build.gradle
@@ -1,0 +1,12 @@
+apply plugin: 'com.android.library'
+
+android {
+    compileSdkVersion 33
+
+    namespace "com.unity3d.player"
+
+    defaultConfig {
+        minSdkVersion 24
+        targetSdkVersion 33
+    }
+}
