+++ Assets/Plugins/Android/mainTemplate.gradle
@@ -5,14 +5,14 @@ apply from: '../shared/keepUnitySymbols.gradle'
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
 // Android Resolver Dependencies Start
-    implementation 'com.google.android.gms:play-services-base:18.7.2' // Assets/Firebase/Editor/DatabaseDependencies.xml:17
-    implementation 'com.google.firebase:firebase-analytics:23.0.0' // Assets/Firebase/Editor/DatabaseDependencies.xml:15
-    implementation 'com.google.firebase:firebase-app-unity:13.1.0' // Assets/Firebase/Editor/AppDependencies.xml:22
-    implementation 'com.google.firebase:firebase-auth:24.0.0' // Assets/Firebase/Editor/AuthDependencies.xml:13
-    implementation 'com.google.firebase:firebase-auth-unity:13.1.0' // Assets/Firebase/Editor/AuthDependencies.xml:20
-    implementation 'com.google.firebase:firebase-common:22.0.0' // Assets/Firebase/Editor/AppDependencies.xml:13
-    implementation 'com.google.firebase:firebase-database:22.0.0' // Assets/Firebase/Editor/DatabaseDependencies.xml:13
-    implementation 'com.google.firebase:firebase-database-unity:13.1.0' // Assets/Firebase/Editor/DatabaseDependencies.xml:22
+    implementation 'com.google.android.gms:play-services-base:18.7.2' // Assets/Imports/Firebase/Editor/DatabaseDependencies.xml:17
+    implementation 'com.google.firebase:firebase-analytics:23.0.0' // Assets/Imports/Firebase/Editor/DatabaseDependencies.xml:15
+    implementation 'com.google.firebase:firebase-app-unity:13.1.0' // Assets/Imports/Firebase/Editor/AppDependencies.xml:22
+    implementation 'com.google.firebase:firebase-auth:24.0.0' // Assets/Imports/Firebase/Editor/AuthDependencies.xml:13
+    implementation 'com.google.firebase:firebase-auth-unity:13.1.0' // Assets/Imports/Firebase/Editor/AuthDependencies.xml:20
+    implementation 'com.google.firebase:firebase-common:22.0.0' // Assets/Imports/Firebase/Editor/AppDependencies.xml:13
+    implementation 'com.google.firebase:firebase-database:22.0.0' // Assets/Imports/Firebase/Editor/DatabaseDependencies.xml:13
+    implementation 'com.google.firebase:firebase-database-unity:13.1.0' // Assets/Imports/Firebase/Editor/DatabaseDependencies.xml:22
     implementation 'com.google.games:gpgs-plugin-support:2.1.0' // Assets/GooglePlayGames/com.google.play.games/Editor/GooglePlayGamesPluginDependencies.xml:11
 // Android Resolver Dependencies End
 **DEPS**}
@@ -42,8 +42,8 @@ android {
     }
 
     defaultConfig {
-        minSdk **MINSDK**
-        targetSdk **TARGETSDK**
+        minSdk 24
+        targetSdk 33
         ndk {
             abiFilters **ABIFILTERS**
             debugSymbolLevel **DEBUGSYMBOLLEVEL**
