+++ Assets/Plugins/iOS/Firebase/libFirebaseCppApp.a.meta
@@ -2,7 +2,7 @@ fileFormatVersion: 2
 guid: 601e8e1d2d2744929ede33676cbeccc0
 labels:
 - gvh
-- gvh_version-13.1.0
+- gvh_version-12.10.1
 - gvhp_exportpath-Plugins/iOS/Firebase/libFirebaseCppApp.a
 timeCreated: 1480838400
 PluginImporter:
