+++ Assets/Plugins/iOS/Firebase/libFirebaseCppAuth.a.meta
@@ -2,7 +2,7 @@ fileFormatVersion: 2
 guid: 44a214618eb6473db34d3808eb040cfb
 labels:
 - gvh
-- gvh_version-13.1.0
+- gvh_version-12.10.1
 - gvhp_exportpath-Plugins/iOS/Firebase/libFirebaseCppAuth.a
 timeCreated: 1480838400
 PluginImporter:
