+++ Assets/Plugins/iOS/Firebase/libFirebaseCppDatabase.a.meta
@@ -2,7 +2,7 @@ fileFormatVersion: 2
 guid: 5a0e503a787b4bd2a2636712f2fb7dc0
 labels:
 - gvh
-- gvh_version-13.1.0
+- gvh_version-12.10.1
 - gvhp_exportpath-Plugins/iOS/Firebase/libFirebaseCppDatabase.a
 timeCreated: 1480838400
 PluginImporter:
