+++ Assets/Plugins/tvOS/Firebase/libFirebaseCppApp.a.meta
@@ -2,7 +2,7 @@ fileFormatVersion: 2
 guid: 46cff64d2cc446bb8fb0137f39fbbf57
 labels:
 - gvh
-- gvh_version-13.1.0
+- gvh_version-12.10.1
 - gvhp_exportpath-Plugins/tvOS/Firebase/libFirebaseCppApp.a
 timeCreated: 1480838400
 PluginImporter:
