+++ Assets/Plugins/tvOS/Firebase/libFirebaseCppAuth.a.meta
@@ -2,7 +2,7 @@ fileFormatVersion: 2
 guid: 255f60ef05484a71a135c57cce9d6fc7
 labels:
 - gvh
-- gvh_version-13.1.0
+- gvh_version-12.10.1
 - gvhp_exportpath-Plugins/tvOS/Firebase/libFirebaseCppAuth.a
 timeCreated: 1480838400
 PluginImporter:
