+++ Assets/Plugins/tvOS/Firebase/libFirebaseCppDatabase.a.meta
@@ -2,7 +2,7 @@ fileFormatVersion: 2
 guid: 67b70764d5594ee5ae00315a1027f697
 labels:
 - gvh
-- gvh_version-13.1.0
+- gvh_version-12.10.1
 - gvhp_exportpath-Plugins/tvOS/Firebase/libFirebaseCppDatabase.a
 timeCreated: 1480838400
 PluginImporter:
