+++ Assets/Resources/PerformanceTestRunSettings.json
@@ -1,0 +1,1 @@
+{"MeasurementCount":-1}
